ScopeVerifController/requirements.txt
absl-py==2.1.0
adbutils==2.8.0
apkutils2==1.0.0
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
cigam==0.0.3
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
ExifRead==3.0.0
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
google-pasta==0.2.0
gpsphoto==2.2.3
grpcio==1.66.2
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
patsy==0.5.6
piexif==1.1.3
pillow==10.4.0
protobuf==4.25.5
py==1.11.0
pyelftools==0.31
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
retry==0.9.2
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
statsmodels==0.14.4
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
xmltodict==0.13.0